requirements.txt
tensorflow==2.15.0
keras==2.15.0
aiogram==3.6.0
numpy==1.26.4
python-dotenv
opencv-python